Refactor into controller and static Module
